requirements.txt
python-dotenv==1.0.1
notebook==7.2.2
ipywidgets==8.1.5
langchain==0.3.7
langchain-community==0.3.7
langchain-huggingface==0.1.2
langchain-ibm==0.3.2
PyMuPDF==1.22.5
pypdf==4.2.0
faiss-cpu==1.9.0
tiktoken==0.8.0
sentence-transformers==3.3.0
chromadb==0.5.0
streamlit==1.24.1
python-dotenv==1.0.0
openai==0.27.8